+++ demo/site/next.config.mjs
@@ -23,6 +23,7 @@ const nextConfig = {
     },
     experimental: {
         optimizePackageImports: ["@comet/site-nextjs"],
+        serverComponentsExternalPackages: ["mjml", "react-intl", "@faire/mjml-react"],
     },
     poweredByHeader: false,
     // https://nextjs.org/docs/advanced-features/security-headers (Content-Security-Policy and CORS are set in middleware/cspHeaders.ts)
@@ -65,8 +66,8 @@ const nextConfig = {
             ],
         },
     ],
-    cacheHandler: process.env.REDIS_ENABLED === "true" ? import.meta.resolve("./dist/cache-handler.js").replace("file://", "") : undefined,
-    cacheMaxMemorySize: process.env.REDIS_ENABLED === "true" ? 0 : undefined, // disable default in-memory caching
+    cacheHandler: process.env.VALKEY_ENABLED === "true" ? import.meta.resolve("./dist/cache-handler.js").replace("file://", "") : undefined,
+    cacheMaxMemorySize: process.env.VALKEY_ENABLED === "true" ? 0 : undefined, // disable default in-memory caching
     rewrites: () => {
         return {
             afterFiles: [